correction, added unique
I realized the file had multiple copies of the same value of TimeStamp.  This caused the PIG program to increase the number of data points from 100,000 to more than 300,000.  And although many of the data points were in exactly the same place I didn't want to add weight to any particular data point, in case a particular interpretation of the data would be influenced by the added weight.